Fix iteration issues
When you delete items from an iterator you need to make sure the
pagination is recalculated properly. Otherwise the page will keep
increasing and you will reach a page with no elements skipping
many items that have not been evaluated.

With this commit we keep deleting only the first page until you get
no results. The filter "created" has been moved to the request to ensure
that GitHub is the one that computes properly the page to be deleted.
The maxWorkflowsToKeep option has been deleted because enters in
conflict with this implementation.